[iOS] refactor: Tap 애니메이션 메소드 리팩토링
- 파일명, 메소드명 변경
- 주석 추가

issue: #99
